split manage controller from user to self, added routes/controller paths
