+++ controllers/users-controller.js
@@ -59,25 +59,8 @@ const userGet = (req, res) => {
     });
 };
 
-const manageUsers = (req, res) => {
-  User
-    .findAll({
-      order: [['name', 'ASC']],
-    //   where: { isEnabled: 0 },
-    })
-    .then((users) => {
-      const statuses = users.map((user) => user.get({ plain: true }));
-      res.render('users/manage-users', { users: statuses });
-    })
-    .catch(({ message } = {}) => {
-      req.log.error(`User.findAll: ${message}`);
-      res.render('error', { message: 'Users not found' });
-    });
-};
-
 module.exports = {
   myProfileGet,
   myProfilePost,
   userGet,
-  manageUsers,
 };
